xxx: update benchmarks
The last benchmarks were taken months ago. It's nice to benchmark the
baseline compiler against the VM and QuickJS.

Initial observations are just as expected: The JIT fares better in
longer, more execution heavy contexts. The VM fares better in terms of
bootup speed. QuickJS absolutely blows us out of the water in terms of
bootup speed, but we blow it out of the water in terms of code
generation optimizations.