ref: vm+stdlib: remove `UnsignedInt` type
This was a long-standing artifact from the interpreter's early days when
it aimed to be a generic VM upon which any interpreted language could
stand on.

Now that we no longer have any such obligations, we can easily get rid
of such PITAs. And we have. This was one of the last ones standing,
simply because how difficult it is to get rid of.