fix: grammar+niche: squash all remaining regressions (that I am aware of)
Ever felt miserable? Try dealing with Niche and you'll think of that
time as a happy one.

It should genuinely be used as a masterclass as to how NOT to implement
an AST to bytecode lowering mechanism. Eh, I designed it as my first
serious lowering mechanism so I guess I'll cut myself some slack.